Identified and removed all fishing communication code from program.cs
